requirements.txt
fastapi==0.95.1
httpx==0.24.0
python-decouple==3.8
typing-extensions==4.5.0
pydantic==1.10.7
uvicorn==0.22.0  # ASGI server to run FastAPI